requirements.txt
aniso8601==9.0.1
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
connexion==2.14.1
Flask==2.3.3
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
flask-swaggerui==0.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
logging==0.4.9.6
MarkupSafe==2.1.5
packaging==24.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
swagger-ui-bundle==0.0.9
swagger-ui-py==23.9.23
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.3.8